Refactor utils structure: Export utils from module index
