+++ src/components/DropDown/DropDown.js
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { Motion, spring } from 'react-motion'
 import ClickOutHandler from 'react-onclickout'
 import theme from '../../theme'
-import { spring as springConf } from '../../utils/styles/spring'
+import { spring as springConf } from '../../utils/styles'
 import { unselectable } from '../../utils/styles'
 import { lerp } from '../../utils/math'
 import getPublicUrl, { styledPublicUrl as asset } from '../../public-url'
