+++ src/components/Header/MenuPanel.js
@@ -4,7 +4,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { Motion, spring } from 'react-motion'
 import ClickOutHandler from 'react-onclickout'
-import { spring as springConf } from '../../utils/styles/spring'
+import { spring as springConf } from '../../utils/styles'
 import { unselectable } from '../../utils/styles'
 import getPublicUrl, { prefixUrl } from '../../public-url'
 import theme from '../../theme'
