+++ src/components/Text/Text.js
@@ -3,7 +3,7 @@ import type { ComponentType, ElementType, Node } from 'react'
 import React from 'react'
 import styled from 'styled-components'
 import theme from '../../theme'
-import { font } from '../../utils/styles/font'
+import { font } from '../../utils/styles'
 
 type Props = {
   children?: Node,
