+++ src/index.js
@@ -11,10 +11,10 @@ export {
 } from 'styled-components'
 
 export { default as theme, themeDark, brand, colors } from './theme'
-export { font } from './utils/styles/font'
-export { grid } from './utils/styles/grid'
-export { spring } from './utils/styles/spring'
-export { breakpoint, BreakPoint } from './utils/styles/breakpoint'
+export { font } from './utils/styles'
+export { grid } from './utils/styles'
+export { spring } from './utils/styles'
+export { breakpoint, BreakPoint } from './utils/styles'
 export { unselectable } from './utils/styles'
 export { default as BaseStyles } from './components/BaseStyles/BaseStyles'
 export { default as Section } from './components/Section/Section'
